split SurfaceHelpers.cpp into GrowSeed.cpp and GrowSegment.cpp
