+++ core/CMakeLists.txt
@@ -35,7 +35,8 @@ target_include_directories(vc_slicing
 
 set(surface_srcs
     src/Surface.cpp
-    src/SurfaceHelpers.cpp
+        src/GrowSeed.cpp
+        src/GrowSurface.cpp
     src/ChunkedTensor.cpp
     src/SurfaceVoxelizer.cpp
 )
