+++ core/src/SurfaceHelpers.hpp
@@ -1,14 +1,29 @@
 #pragma once
 
-#include <opencv2/core.hpp>
+#include <iomanip>
 
-// External declarations for configurable parameters
-extern float local_cost_inl_th;
-extern float same_surface_th;
-extern float straight_weight;
-extern float straight_min_count;
-extern int inlier_base_threshold;
 
-cv::Mat_<cv::Vec3f> upsample_with_grounding(cv::Mat_<cv::Vec3f> &small, cv::Mat_<cv::Vec2f> &locs, const cv::Size &tgt_size, const cv::Mat_<cv::Vec3f> &points, double sx, double sy);
-void refine_normal(const std::vector<std::pair<cv::Vec2i,cv::Vec3f>> &refs, cv::Vec3f &point, cv::Vec3f &normal, cv::Vec3f &vx, cv::Vec3f &vy, const std::vector<float> &ws);
-std::string get_surface_time_str();
+static inline std::string get_surface_time_str()
+{
+    using namespace std::chrono;
+
+    // get current time
+    auto now = system_clock::now();
+
+    // get number of milliseconds for the current second
+    // (remainder after division into seconds)
+    auto ms = duration_cast<milliseconds>(now.time_since_epoch()) % 1000;
+
+    // convert to std::time_t in order to convert to std::tm (broken time)
+    auto timer = system_clock::to_time_t(now);
+
+    // convert to broken time
+    std::tm bt = *std::localtime(&timer);
+
+    std::ostringstream oss;
+
+    oss << std::put_time(&bt, "%Y%m%d%H%M%S"); // HH:MM:SS
+    oss << std::setfill('0') << std::setw(3) << ms.count();
+
+    return oss.str();
+}
